Reformat XML file
Now going to use the <last_updated> element as a way of checking the last USER to create a memo - rather than the last time the file was updated.

This will allow the parser to go through users and then find the latest memo ID.

Alternately, we could use the <last_updated> to store the next ID for a new memo.