Remove write priveleges from guest
The theme behind this "Memo Database" is something like a government memo serivce. Every user on the system can create a memo and post it for others to see.

However, for the regular "joes", they are only allowed to browse the creates memos, not create them. Regulars will use the guest account.