Ensure latest memo displayed on reload
The body onload function now passes in a parameter to the showmemo function each time it's called. This means that we should be 100% certain the latest memo is displayed on page refresh/redirects.

If the function is called with reloaded set to false, it will still cycle through the memos.